Refactor module loader to operate async
Loader now returns and EventEmitter that can emit and event with data
for each loaded module, as well as a final event with all modules.

This update required major changes being made to the primary BentoBox api.
The exported module is now an object factory, allowing its configuration to
be loaded asynchonously.